Add more rendering tests
I'm starting to think this is not a great idea. The parser seem to have
some werid things going on and I am not confident I can get it to work
without it fucking up the documents every now and then. At least not
without doing a major rewrite of the library.